+++ Changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0] - 2019-03-05
+- Add countLines for UiTextarea, that depend with height
+- Add required styles for UiTextarea
+- Fix UiTextarea styles for border and margin
+
+## [1.12.0] - 2019-03-05
+- Add required styles for Select
+
 ## [1.11.0] - 2019-03-04
 - Add style for required field in TextField
 - Fix empty field for DateInput and DateTimeInput